Split omemo plugin into more files
