+++ src/plugins/omemo/consts.js
@@ -1,0 +1,3 @@
+export const UNDECIDED = 0;
+export const TRUSTED = 1;
+export const UNTRUSTED = -1;
