+++ src/plugins/omemo/devices.js
@@ -1,0 +1,4 @@
+import Device from './device.js';
+import { Collection } from '@converse/skeletor/src/collection';
+
+export default Collection.extend({ model: Device });
