+++ src/plugins/omemo/errors.js
@@ -1,0 +1,7 @@
+export class IQError extends Error {
+    constructor (message, iq) {
+        super(message, iq);
+        this.name = 'IQError';
+        this.iq = iq;
+    }
+}
